+++ examples/testfull/swift-testfull/colobject.h
@@ -1,0 +1,36 @@
+//
+//  colobject.h
+//  swift-testfull
+//
+//  Created by Sébastien Keim on 18/02/2024.
+//
+
+#ifndef colobject_h
+#define colobject_h
+
+/*
+ external objects doesn't work for the moment
+ 
+ this stub allow the binding to compile
+ 
+ */
+
+
+
+typedef enum  {
+      make_upper,
+      make_capitalize,
+      make_lower,
+} Transformer_t;
+
+
+
+
+typedef enum  {
+      ZERO,
+      ONE
+} Point_t;
+
+
+#endif /* colobject_h */
+