+++ examples/testfull/swift-testfull/main.swift
@@ -1,0 +1,11 @@
+//
+//  main.swift
+//  swift-testfull
+//
+ 
+import Foundation
+
+let ONE = 1
+
+innerjoin()
+strencoding()