+++ src/main.rs
@@ -21,6 +21,7 @@ mod config;
 mod index;
 mod langc;
 mod langrust;
+mod langswift;
 mod language;
 mod lint;
 mod log;