+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ## insrcdata 0.3.0 (not published)
 
 ### backward-incompatible changes
+* non_exhaustive flag in rust for label columns
+
 ### other
 * help fields for doc comment generation
-
+* swift language support as binding to C generated code. The Object column type is not available for this language.
 
 ## insrcdata 0.2.0 (2023-10-27)
 Publish new version for build.rs support